refactor(pm): move DirFd from util/ to fs/ submodule
`DirFd` is a filesystem primitive, so it fits better alongside the
other fs operations the project already re-exports. Turn `fs.rs` into
`fs/mod.rs` and place `at.rs` as a Linux-only submodule.

Call sites update from `crate::util::at::DirFd` to `crate::fs::at::DirFd`.
