+++ crates/pm/src/util/mod.rs
@@ -1,5 +1,3 @@
-#[cfg(target_os = "linux")]
-pub mod at;
 pub mod binary_resolver;
 pub mod cache;
 pub mod cli_enum;
